% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lacunarity.R
\name{genlac}
\alias{genlac}
\title{Generalized lacunarity index}
\usage{
genlac(x)
}
\arguments{
\item{x}{is a vector of 0's and 1's}
}
\value{
s is the box scale

q is the moment order

y(q) is the lacunarity scaling exponent gamma(q)

Dqs is the lacunarity for the box scale and q-moment
}
\description{
Calculates the Generalized lacunarity index of a binary series
}
\examples{
x<-rbinom(1000,1,0.85)
genlac(x)
}
