% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lacunarity.R
\name{lac}
\alias{lac}
\title{lacunarity index}
\usage{
lac(x)
}
\arguments{
\item{x}{is a vector of 0's and 1's}
}
\value{
y is the lacunarity scaling exponent

Ds is the lacunarity for the box scale

s is the box scale
}
\description{
Calculates the lacunarity index of a binary series
}
\examples{
x<-rbinom(1000,1,0.85)
lac(x)
}
