% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lacunarity.R
\docType{package}
\name{lacunarity-package}
\alias{lacunarity}
\alias{lacunarity-package}
\title{lacunarity: lacunarity and generalized lacunarity for binary time series}
\description{
Estimators of the gliding-box lacunarity index and of the generalized
(multifractal-like) lacunarity for one-dimensional binary series. See the
package vignette, \code{vignette("lacunarity")}, for the underlying theory
and worked examples.
}
\section{Main functions}{

\describe{
  \item{\code{\link{lac}}}{ordinary lacunarity index \eqn{\Lambda(s)} and its
    scaling exponent.}
  \item{\code{\link{genlac}}}{generalized lacunarity \eqn{\Lambda_q(s)} and
    the spectrum of exponents \eqn{\gamma(q)}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ikarobarreto/lacunarity}
  \item Report bugs at \url{https://github.com/Ikarobarreto/lacunarity/issues}
}

}
\author{
\strong{Maintainer}: Ikaro Barreto \email{daniel.carvalho.ib@gmail.com}

}
\keyword{internal}
