% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lacunarity.R
\name{zqs}
\alias{zqs}
\title{Moments generating Function}
\usage{
zqs(mat, q)
}
\arguments{
\item{mat}{is a two-column matrix with x and p(x)}

\item{q}{is the moment order}
}
\value{
q-moment of a frequency distribution of the box masses
}
\description{
Calculates the q-moment of frequency distribution of the box masses
}
